Show warning when adding a co-signer
